content/canvas/test/test_2d.composite.operation.unrecognised.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.composite.operation.unrecognised - bug 401788</title>
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var _thrown_outer = false;
try {

ctx.globalCompositeOperation = 'xor';
ctx.globalCompositeOperation = 'nonexistent';
ok(ctx.globalCompositeOperation == 'xor', "ctx.globalCompositeOperation == 'xor'");

} catch (e) {
    _thrown_outer = true;
}
todo(!_thrown_outer, 'should not throw exception');

SimpleTest.finish();

});
</script>
